FY-156/feat: GamePlaying 컴포넌트 삭제 및 PongScenePage, BracketPage 컴포넌트 신설
